All changes for version 1.1.0
