+++ META-INF/MANIFEST.MF
@@ -1,5 +1,5 @@
 Manifest-Version: 1.0
 Bundle-ManifestVersion: 2
 Bundle-Name: Web
-Bundle-SymbolicName: com.siemens.bt.jazz.workitemeditor.gitLabConnector;singleton:=true
-Bundle-Version: 1.0.0.qualifier
+Bundle-SymbolicName: com.siemens.bt.jazz.workitemeditor.gitCommitPicker;singleton:=true
+Bundle-Version: 1.1.0.qualifier
