--- build.properties
@@ -1,4 +1,0 @@
-bin.includes = plugin.xml,\
-               META-INF/,\
-               *.xml,\
-               resources/
