--- build.xml
@@ -1,131 +1,0 @@
-<project name="GitLabConnector" default="dist" basedir=".">
-	<description>GitLabConector build script</description>
-	<!-- define the alias of your destination application; e.g. ccm, jts, rm, qm-->
-	<property name="appType" value="ccm" />
-	<!-- name of the update-site folder -->
-	<property name="page-extension-site" value="com.siemens.bt.jazz.workitemeditor.gitLabConnector_updatesite" />
-	<!-- identifier of the plugin -->
-	<property name="pluginID" value="com.siemens.bt.jazz.workitemeditor.gitLabConnector" />
-	<!-- name of the Zip file -->
-    <!-- Set build label depending on running on the jbe or not. -->
-    <condition property="fileName" value="GitLabConnector_${buildLabel}" else="GitLabConnector">
-        <isset property="buildLabel" />
-    </condition>
-	<!-- identifier of the feature -->
-	<property name="featureID" value="${pluginID}.feature" />
-	<!-- current version -->
-	<property name="version" value="1.0.0" />
-	<!-- path to node installation -->
-	<property name="projectSource" value="${pluginID}" />
-	<property name="features" value="features" />
-	<property name="plugins" value="plugins" />
-	<property name="this.dir" location="." />
-	<property name="dist.dir" location="${this.dir}/dist" />
-	<property name="home.dir" location="${dist.dir}/${page-extension-site}" />
-	<property name="feature.dir" location="${home.dir}/${features}" />
-	<property name="plugins.dir" location="${home.dir}/${plugins}" />
-	<property name="deployment.props" location="${this.dir}/deployment-properties.ini" />
-
-	<target name="init">
-		<tstamp />
-		<mkdir dir="${home.dir}" />
-		<mkdir dir="${feature.dir}" />
-		<mkdir dir="${plugins.dir}" />
-	</target>
-
-	<target name="createPropertyFile" depends="init">
-		<propertyfile file="${deployment.props}">
-			<entry key="pluginFolderName" value="${page-extension-site}"/>
-			<entry key="pluginFeatureId" value="${featureID}"/>
-		</propertyfile>
-	</target>
-
-	<target name="createInIfile" depends="init">
-		<propertyfile file="${dist.dir}/${page-extension-site}.ini" comment="Copy this file into conf/${appType}/provision_profiles folder">
-			<entry key="url" value="file:${appType}/sites/${page-extension-site}"/>
-			<entry key="featureid" value="${featureID}"/>
-		</propertyfile>
-		<replace file="${dist.dir}/${page-extension-site}.ini" token="\:" value=":"/>
-	</target>
-
-	<target name="packPlugin">
-		<copy todir="${plugins.dir}">
-			<fileset dir="${this.dir}">
-		 		<exclude name="build.*"/>
-		 		<exclude name="*.zip"/>
-		 		<exclude name=".settings"/>
-		 		<exclude name="conf/"/>
-		 		<exclude name="debug-setup/"/>
-		 		<exclude name="dist/"/>
-		 		<exclude name="launch/"/>
-			</fileset>
-		</copy>
-		<zip destfile="${plugins.dir}/${pluginID}_${version}.${DSTAMP}${TSTAMP}.jar" basedir="${plugins.dir}" />
-		<delete includeemptydirs="true">
-		    <fileset dir="${plugins.dir}">
-		    	<exclude name="*.jar"/>
-		    </fileset>
-		</delete>
-	</target>
-
-	<target name="createFeatureXML">
-		<echoxml file="${feature.dir}/feature.xml">
-			<feature 
-				id= "${featureID}"
-				label= "Feature"
-				version="${version}.${DSTAMP}${TSTAMP}"
-				provider-name= "SIEMENS BUILDING TECHNOLOGIES">
-				
-				<description>
-                      Connects gitLab commits with a "RTC Work Item"
-				</description>
-				<copyright>
-				      Copyright (c) 2017, Siemens AG
-				</copyright>
-				<license>
-                      MIT License
-				</license>
-				<requires>
-					<import plugin="net.jazz.ajax" version="2.3.2" match="greaterOrEqual"/>
-					<import plugin="com.ibm.team.workitem.web" version="3.1.1100" match="greaterOrEqual"/>
-				</requires>
-				<plugin
-				         id="${pluginID}"
-				         download-size="0"
-				         install-size="0"
-				         version="${version}.${DSTAMP}${TSTAMP}"/>
-			</feature>
-		</echoxml>
-	</target>
-
-	<target name="createSiteXML">
-		<echoxml file="${dist.dir}/${page-extension-site}/site.xml">
-			<site>
-				<feature 
-					url="features/${featureID}_${version}.${DSTAMP}${TSTAMP}.jar" 
-					id="${featureID}" 
-					version="${version}.${DSTAMP}${TSTAMP}" />
-			</site>
-		</echoxml>
-		<echo message="created site.xml" />
-	</target>
-
-	<target name="createFeatureJar" depends="createSiteXML">
-		<echo message= "creating feature jar" />
-		<zip destfile="${feature.dir}/${featureID}_${version}.${DSTAMP}${TSTAMP}.jar">
-			<fileset dir="${feature.dir}" />
-		</zip>
-	</target>
-
-	<target name="dist" depends="createInIfile,createPropertyFile,createFeatureXML,packPlugin,createFeatureJar">
-		<delete file="${feature.dir}/feature.xml" />
-		<zip destfile="${fileName}.zip">
-			<zipfileset dir="${dist.dir}" />
-		</zip>
-		<echo message="dist.zip created" />
-		<delete includeemptydirs="true">
-		    <fileset dir="${dist.dir}" />
-		</delete>
-		<echo message="deleted used files" />
-	</target>
-</project>
