+++ deployment-properties.ini
@@ -1,0 +1,2 @@
+pluginFeatureId=com.siemens.bt.jazz.workitemeditor.gitCommitPicker.feature
+pluginFolderName=com.siemens.bt.jazz.workitemeditor.gitCommitPicker_updatesite
