--- resources/ui/widget/templates/ConnectCommitsTemplate.html
@@ -1,5 +1,0 @@
-<div>
-    <div class="gitLab-connect-commits">
-		<div class="repoPane" style="margin:20px;"></div>
-    </div>
-</div>
