+++ resources/ui/widget/templates/PickCommitsTemplate.html
@@ -1,0 +1,5 @@
+<div>
+    <div class="gitLab-connect-commits">
+		<div class="repoPane" style="margin:20px;"></div>
+    </div>
+</div>
